refactor(core): remove Android XPath inspection API
