+++ packages/core/tests/unit-test/task-runner/index.test.ts
@@ -149,12 +149,6 @@ describe(
       const uiTree: UITreeSnapshot = {
         platform: 'android',
         capturedAt: 1,
-        xpathPolicy: {
-          stableAttrs: ['resource-id'],
-          textAttrs: ['content-desc', 'text'],
-          excludedTargetTypes: [],
-          max: 5,
-        },
         root: {
           type: 'Window',
           attrs: {},
@@ -240,12 +234,6 @@ describe(
       const uiTree: UITreeSnapshot = {
         platform: 'android',
         capturedAt: 1,
-        xpathPolicy: {
-          stableAttrs: ['resource-id'],
-          textAttrs: ['content-desc', 'text'],
-          excludedTargetTypes: [],
-          max: 5,
-        },
         root: {
           type: 'Window',
           attrs: {},
